changes to sg_v40.html for removal of master-slave
A new Linux kernel directive has outlawed the master/slave
terminology in "new" code. Take the ooportunity to replace them
with more precide terms: read-side and write-side. That loses
the sense that the read-side is the "leader" and the write-side
is the "follower" (and leader/follower are approved replacement
terms). The newly reworked diagrams in this page should make
that latter point clear.