beta sg3_utils-1.46 revision 856; new sgv4 tarball
The changes here are many and varied and stem from the Linux
code style directive to remove the use of master/slave. This
impacts the newer parts of the sgv4 driver including the
names of some (extended) ioctls. It also impacts some of
some of the sg3_utils utilities in its testing directory.