announce release of lsscsi-0.32
This corresponds to the author's subversion revision 167 of lsscsi.
Looks like Fedora 34 has jumped the gun with a rpm claiming to be
release 0.32 that looks like svn revision 164. The author will try
to make this clearer in the future by placing _pre_-release in
the ChangeLog and version string shown by 'lsscsi --version'.